+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 0.3.15
+## 0.4.0--alpha.1
+
+- [helpers] use generics when typing `filterKeysFromObj`
+- [slider] delete exported css
+- [slider] better document props
+- [slider] re-export slide context
+- [field] re-export field context
+- [form] re-export form context
+
+## 0.4.0--alpha.0
 
 - [helpers] add some lodash-esque functions
 - [helpers] fix types after upgrading typescript